Create BoundedFeltExpr as a VarWrapper

+++ crates/air_infra/src/core/expressions/felt_expr.rs
@@ -120,7 +120,7 @@ impl Serialize for FeltExpr {
 macro_rules! const_expr {
     ($val:expr) => {
         FeltExpr::Var($crate::core::expressions::var_expr::VarExpr::new_const(
-            $crate::core::Felt::from($val),
+            $crate::core::Felt::from($val as u32),
         ))
     };
 }
